Netdump example changed for querying process
The previous version of the netdump example used FilterFlags.Sniff,
which produces hit-and-miss results when querying the process behind the
packet, because the process may be gone or unbound from the port
reported in the packet header by the time the query is made. See the
comments added to the Program.cs for more.